web-socket-server/tests/pyproject.toml
[tool.poetry]
name = "web-socket-server_tests"
version = "0.1.0"
description = ""
authors = ["hamadakafu <kafu.h1998@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-socketio = "*"
requests = "*"

[tool.poetry.dev-dependencies]
pytest = "^5.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
